Require cytoolz (#45)
* Require cytoolz

While cytoolz is an optional dependency of distributed and can be subbed
in for toolz, it is nice to include as it can provide some speedups over
toolz. As these are conda packages, it is pretty easy to pull in this
package as well.

* Bump build number to 1

As we now include cytoolz as a dependency, go ahead and bump the build
number to produce a new package with the cytoolz dependency.